HttpSys request headers Keys and Count both allocate
RequestHeaders's Key and Count with new implementation. Count is allocation free, while Keys allocate the string for the header keys that are not known and a list for the result.
RequestHeader.Generated.cs has a new HeaderKeys returned as ReadOnlySpan<byte> for accessing the know keys - by avoiding allocation with enumerator or a static list.
RequestHeader.Generated.tt is updated, but I don't see it being used to generate the RequestHeader.Generated.cs file.
NativeRequestContext provides new implementation to access the header keys and count.